codenation_dev-ps/requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
entrypoints==0.3
flake8==3.7.7
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
python-dotenv==0.10.2
requests==2.22.0
rope==0.14.0
six==1.12.0
typed-ast==1.3.5
urllib3==1.25.3
wrapt==1.11.1
